feat: share curated video profiles with MCP planning
