+++ docs/research/2026-08-24-video-capability-shared-layer.md
@@ -46,6 +46,19 @@
 
 ## 共享层应该保存什么
 
+### 共享层接入前的 UI 上下文闸
+
+GUI 已经能让用户选择的模型、模式、参数、参考槽、变体和供应商覆盖，属于现有产品事实，不是只供界面显示的装饰信息。MCP/Agent 在做推荐或规划前必须先读取同一个档案和当前 catalog/mapping 上下文，再把用户自然语言、素材和目标投影进去；禁止在 MCP 内另造一套简化模型列表、参数枚举或参考角色。
+
+对照检查固定为：
+
+1. GUI 从哪里读取模型档案、当前变体、模式和参数控件；
+2. GUI 如何把参考素材按槽位和角色组装成真实请求；
+3. 当前供应商 mapping 如何消费这些字段；
+4. MCP 是否读取同一 owner，并在用户切换模型/模式/素材/参数后重新计算，而不是沿用上一次推荐。
+
+如果 GUI 已有事实但 MCP 尚未接通，优先打通读取边界；只有 UI 与现有接入都没有的事实才进入新设计。这样用户在 GUI 和 MCP 之间切换时看到的是同一项能力，差异只来自当前入口的表达方式，不来自两套模型知识。
+
 共享层只保存四类稳定原语，全部按 `provider + model + version + mode` 作用域声明：
 
 ### 1. 输入角色（Input roles）
@@ -122,6 +135,8 @@ type ExpressionChannel = {
 5. 2.5 的 APIMart 页面本轮抓取返回 404。当前仓库已有来源记录，但在把它作为共享注册表的生产事实前，必须重新打开可访问的官方原文或通过实际 APIMart schema 重新核验；在此之前状态保持 `unknown`，不扩大承诺。
 
 6. 默认候选不再写死一份“我们以为供应商有的模型列表”。Electron/stdio 启动时从用户当前模型目录读取视频模型：命中已对账档案才使用精确模式、槽位和参数；目录里出现但尚未对账的模型会生成保守的 `unknown` 档案，只提供文生/单图两条可证明的基础入口，不凭空提供首尾帧、全能参考或原生运镜。这样模型缺少某个高级能力时不会把整项视频能力关掉，供应商新增模型也不需要改推荐器。
+
+7. UI↔MCP 上下文一致性是独立验收项：每个共享候选都必须能追溯到 GUI 的 `archetypeId`、当前 catalog 行和 mapping；模型切换、模式切换、参考替换和参数编辑都要在两入口得到同样的可用范围与请求字段，不能只验证 MCP 自己的纯逻辑。
 
 本轮官方 APIMart 视频目录能检索到 `doubao-seedance-2-0`，但没有可访问的 Seedance 2.5 详细页面；因此 2.5 档案保留给明确出现在用户目录的模型，不能作为默认推荐候选。
