+++ electron/capabilityCore/mcpStdioServer.ts
@@ -39,7 +39,7 @@ import type { ModuleRegistry } from './moduleRegistry'
 import { createCatalogModuleRegistry } from './moduleCatalogBootstrap'
 import { createGenerationProviderBootstrap } from './generationProviderBootstrap'
 import { readCatalog } from '../catalog/catalogStore'
-import { buildVideoModelCandidates, recommendVideoGeneration } from '../shared/videoCapabilities'
+import { buildVideoModelCandidates, recommendVideoGeneration, videoArchetypeIdFromMeta } from '../shared/videoCapabilities'
 
 const productionRuns = getProductionRunService()
 
@@ -200,6 +200,7 @@ export async function startMcpStdioServer(authorities: McpStdioServerOptions = {
       provider: model.vendorKey,
       modelKey: model.modelKey,
       label: model.labelZh,
+      archetypeId: videoArchetypeIdFromMeta(model.meta),
       parameterControls: model.onboarding?.fields?.map((field) => ({
         key: field.key,
         label: field.displayName,
